Start of data management refactor
