+++ apps/stakes/package.json
@@ -38,7 +38,8 @@
     "xstate": "^4.7.0-rc6",
     "types": "~1.0.0",
     "trading-data": "~1.0.1",
-    "just-safe-get": "~2.0.0"
+    "just-safe-get": "~2.0.0",
+    "@sveltestack/svelte-query": "~1.0.6"
   },
   "devDependencies": {
     "@babel/core": "~7.11.6",
@@ -71,15 +72,15 @@
     "postcss-cli": "^6.1.3",
     "postcss-import": "^12.0.1",
     "postcss-url": "^8.0.0",
-    "rollup": "~2.27.1",
+    "rollup": "~2.38.5",
     "rollup-plugin-postcss": "~3.1.8",
-    "rollup-plugin-svelte": "~6.0.1",
-    "rollup-plugin-terser": "^4.0.4",
-    "sapper": "~0.28.7",
-    "svelte": "~3.28.0",
+    "rollup-plugin-svelte": "~7.1.0",
+    "rollup-plugin-terser": "~7.0.2",
+    "sapper": "~0.29.1",
+    "svelte": "~3.32.1",
     "svelte-awesome": "^2.2.1",
     "svelte-notifications": "^0.9.8",
-    "svelte-preprocess": "~4.3.0",
+    "svelte-preprocess": "~4.6.6",
     "svelte-preprocess-postcss": "^1.1.1",
     "svelte-select": "^3.1.2",
     "tailwindcss": "~1.8.10",
