--- apps/stakes/src/css.js
@@ -1,8 +1,0 @@
-const camelCaseToDash = (str) =>
-  str.replace(/([a-zA-Z])(?=[A-Z])/g, '$1-').toLowerCase();
-
-export default (props) =>
-  Object.keys(props).reduce(
-    (str, key) => `${str}; ${camelCaseToDash(key)}: ${props[key]}`,
-    ''
-  );
